requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
python-multipart
google-generativeai
google-cloud-aiplatform